consolidated into 1 test file
